refactor: qc-core post 도메인 패키지 정리
