+++ qc-application/src/main/java/com/eager/questioncloud/application/api/workspace/CreatorPostService.java
@@ -1,8 +1,8 @@
 package com.eager.questioncloud.application.api.workspace;
 
 import com.eager.questioncloud.core.common.PagingInformation;
-import com.eager.questioncloud.core.domain.post.PostListItem;
-import com.eager.questioncloud.core.domain.post.PostRepository;
+import com.eager.questioncloud.core.domain.post.dto.PostListItem;
+import com.eager.questioncloud.core.domain.post.infrastructure.PostRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
