+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/post/infrastructure/PostCommentEntity.java
@@ -1,5 +1,6 @@
-package com.eager.questioncloud.core.domain.post;
+package com.eager.questioncloud.core.domain.post.infrastructure;
 
+import com.eager.questioncloud.core.domain.post.model.PostComment;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
