+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/post/infrastructure/PostCommentRepository.java
@@ -1,6 +1,8 @@
-package com.eager.questioncloud.core.domain.post;
+package com.eager.questioncloud.core.domain.post.infrastructure;
 
 import com.eager.questioncloud.core.common.PagingInformation;
+import com.eager.questioncloud.core.domain.post.dto.PostCommentDetail;
+import com.eager.questioncloud.core.domain.post.model.PostComment;
 import java.util.List;
 
 public interface PostCommentRepository {
