+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/post/infrastructure/PostJpaRepository.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.core.domain.post;
+package com.eager.questioncloud.core.domain.post.infrastructure;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
