+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/post/model/Post.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.core.domain.post;
+package com.eager.questioncloud.core.domain.post.model;
 
 import java.time.LocalDateTime;
 import lombok.Builder;
