+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/post/model/PostFile.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.core.domain.post;
+package com.eager.questioncloud.core.domain.post.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
